+++ src/main/java/uk/ac/ox/oxfish/biology/LocalBiology.java
@@ -19,7 +19,7 @@ public interface LocalBiology
      * @param specie  the specie you care about
      * @return the biomass of this specie
      */
-    public Integer getBiomass(Specie specie);
+    public Double getBiomass(Specie specie);
 
 
     /**
@@ -27,7 +27,7 @@ public interface LocalBiology
      * @param specie the specie fished
      * @param biomassFished the biomass fished
      */
-    public void reactToThisAmountOfBiomassBeingFished(Specie specie, Integer biomassFished);
+    public void reactToThisAmountOfBiomassBeingFished(Specie specie, Double biomassFished);
 
 
 }
